src/dashboard/simple-dashboard-tests.py
#!/usr/bin/env python3
"""
Simple Dashboard Tests - Basic functionality testing without Playwright
"""

import requests
import time
import json
from datetime import datetime

class SimpleDashboardTests:
    def __init__(self):
        self.dashboard_url = "http://localhost:5003"
        self.test_results = []
    
    def test_dashboard_responds(self):
        """Test that dashboard responds to HTTP requests"""
        try:
            response = requests.get(self.dashboard_url, timeout=10)
            assert response.status_code == 200, f"Expected 200, got {response.status_code}"
            
            content = response.text
            assert "AI Development Team Dashboard" in content, "Dashboard title not found"
            assert "Overview" in content, "Overview tab not found"
            assert "Agents" in content, "Agents tab not found"
            assert "System" in content, "System tab not found"
            assert "Pipeline" in content, "Pipeline tab not found"
            
            print("✅ Dashboard responds correctly")
            return True
            
        except Exception as e:
            print(f"❌ Dashboard response test failed: {e}")
            return False
    
    def test_api_endpoints(self):
        """Test dashboard API endpoints"""
        try:
            # Test data endpoint
            response = requests.get(f"{self.dashboard_url}/api/data", timeout=5)
            if response.status_code == 200:
                data = response.json()
                assert "timestamp" in data, "Timestamp not in API response"
                print("✅ API endpoints working")
                return True
            else:
                print("⚠️ API endpoint not available (may not be implemented)")
                return True  # Not critical
                
        except Exception as e:
            print(f"⚠️ API test warning: {e}")
            return True  # Not critical
    
    def test_static_content(self):
        """Test that static content loads"""
        try:
            response = requests.get(self.dashboard_url, timeout=10)
            content = response.text
            
            # Check for required elements
            required_elements = [
                "tab-content",
                "metric-card", 
                "agent-card",
                "progress-bar",
                "socket.io"
            ]
            
            for element in required_elements:
                assert element in content, f"Required element not found: {element}"
            
            print("✅ Static content loads correctly")
            return True
            
        except Exception as e:
            print(f"❌ Static content test failed: {e}")
            return False
    
    def test_websocket_connection(self):
        """Test WebSocket connectivity (basic check)"""
        try:
            import socketio
            
            sio = socketio.SimpleClient()
            sio.connect(self.dashboard_url, timeout=5)
            
            # Emit a test event
            sio.emit('request_update')
            
            # Wait for response
            time.sleep(2)
            
            sio.disconnect()
            print("✅ WebSocket connection working")
            return True
            
        except ImportError:
            print("⚠️ WebSocket test skipped (socketio not available)")
            return True
        except Exception as e:
            print(f"⚠️ WebSocket test warning: {e}")
            return True  # Not critical for basic functionality
    
    def test_agent_data_integration(self):
        """Test that dashboard can read agent data"""
        try:
            # Check if simulation files exist
            import os
            
            expected_files = [
                "agent_simulation_state.json",
                "agent_activity.log"
            ]
            
            files_found = 0
            for filename in expected_files:
                if os.path.exists(filename):
                    files_found += 1
                    print(f"✅ Found: {filename}")
            
            if files_found > 0:
                print("✅ Agent data integration working")
                return True
            else:
                print("⚠️ No agent data files found (run agent-activity-simulator.py)")
                return True  # Not critical
                
        except Exception as e:
            print(f"⚠️ Agent data test warning: {e}")
            return True
    
    def test_performance(self):
        """Test dashboard performance"""
        try:
            start_time = time.time()
            response = requests.get(self.dashboard_url, timeout=10)
            end_time = time.time()
            
            response_time = end_time - start_time
            
            assert response_time < 5.0, f"Response too slow: {response_time:.2f}s"
            assert len(response.content) > 1000, "Response content too small"
            
            print(f"✅ Performance test passed ({response_time:.2f}s)")
            return True
            
        except Exception as e:
            print(f"❌ Performance test failed: {e}")
            return False
    
    def run_all_tests(self):
        """Run all dashboard tests"""
        print("🧪 SIMPLE DASHBOARD TESTS")
        print("=" * 50)
        print(f"Dashboard URL: {self.dashboard_url}")
        print(f"Test time: {datetime.now().isoformat()}")
        print("=" * 50)
        
        tests = [
            ("Dashboard Responds", self.test_dashboard_responds),
            ("API Endpoints", self.test_api_endpoints),
            ("Static Content", self.test_static_content),
            ("WebSocket Connection", self.test_websocket_connection),
            ("Agent Data Integration", self.test_agent_data_integration),
            ("Performance", self.test_performance)
        ]
        
        results = {}
        for test_name, test_func in tests:
            print(f"\n🔄 Running: {test_name}")
            try:
                result = test_func()
                results[test_name] = result
            except Exception as e:
                print(f"❌ {test_name} failed with exception: {e}")
                results[test_name] = False
        
        # Summary
        print("\n" + "=" * 50)
        print("📊 TEST RESULTS SUMMARY")
        print("=" * 50)
        
        passed = sum(1 for r in results.values() if r)
        total = len(results)
        
        for test_name, result in results.items():
            status = "✅ PASS" if result else "❌ FAIL"
            print(f"{status} {test_name}")
        
        success_rate = (passed / total) * 100
        print(f"\n📈 Overall: {passed}/{total} tests passed ({success_rate:.1f}%)")
        
        if success_rate >= 80:
            print("🎉 Dashboard is working well!")
        elif success_rate >= 60:
            print("⚠️ Dashboard has some issues but is functional")
        else:
            print("❌ Dashboard needs significant fixes")
        
        return results

def main():
    """Main test runner"""
    # Check if dashboard is running
    tester = SimpleDashboardTests()
    
    print("🔍 Checking dashboard availability...")
    try:
        response = requests.get(tester.dashboard_url, timeout=5)
        if response.status_code == 200:
            print("✅ Dashboard is running")
            results = tester.run_all_tests()
            
            # Save results
            with open("dashboard_test_results.json", 'w') as f:
                json.dump({
                    "timestamp": datetime.now().isoformat(),
                    "results": results,
                    "dashboard_url": tester.dashboard_url
                }, f, indent=2)
            
            print(f"\n📄 Results saved to dashboard_test_results.json")
            
        else:
            print(f"❌ Dashboard returned status {response.status_code}")
            
    except requests.exceptions.ConnectionError:
        print("❌ Cannot connect to dashboard")
        print("💡 Start dashboard: python working-dashboard.py")
    except Exception as e:
        print(f"❌ Error: {e}")

if __name__ == "__main__":
    main()
