requirements-dashboard.txt
# VF Live Dashboard Requirements
flask==2.3.3
flask-socketio==5.3.6
boto3==1.34.84
botocore==1.34.84
psutil==5.9.5
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3
gunicorn==21.2.0
requests==2.31.0
